requirements.txt
pandas>=2.0.0
PyYAML>=6.0.0
yfinance>=0.2.0
requests>=2.31.0
